+++ .travis/setup-pypostgres.sh
@@ -1,2 +1,0 @@
-pip install -U pg8000
-`dirname $0`/postgres.sh